Show confirmation dialog if there are multiple open tabs on closing
As a followup #149 this fixes the behavior of the code to match the
description of the respective gsettings entry "confirm-window-close".
Said entry is also updated to reflect the change added in #149.
No changes to the gsettings handling were made thus only users
who have "confirm-window-close" turned on already will see the
new behavior.