terminal-screen: Allow passing of right click to applications
Currently, the popup menu is shown for all right clicks that don't have
Ctrl, Alt, or Shift.  This is the opposite of what is done in every
other terminal.  The right click should go to the underlying application
except when a special modifier is used.  Gnome utilizes Shift as the
modifier and LXDE uses Ctrl as their modifier.

Referencing gnome-terminal-3.40:
https://github.com/GNOME/gnome-terminal/blob/gnome-3-40/src/terminal-screen.c#L1779C1-L1811C16 we can port this logic over and then we achieve exactly
what we are looking for. If there is no application, the terminal
handles the right click normally.  If there is an application, the click
gets passed through unless Shift is held.

Closes #388